Load the dictionary from https://skk-dev.github.io/dict/
The dictionaries are not available at http://skk-dict-mirror.appspot.com/ any
more.  This change replaces the URL with the ones at
https://skk-dev.github.io/dict.  This is a temporary change and is going to be
configurable so that the user can specify any URLs as they like.

The change also introduces a third party library, pako, for deflating a gzip
file.  This is just a hacky way to introduce a library into the extension and
we need to take advantage of any package system to install the library.