Fix press on the Ctrl key commits the conversion
For some reasons, a hack to determine whether a pressed key is a
modifier key or not no longer works.

```
keyevent.key.charCodeAt(0) == 0xFFFD
```

Instead, we need to call `keyevent.ctrlKey && keyevent.key == "Ctrl"` to
make sure it is the Ctrl key.  This change also makes sure that pressing
on the Alt key also is ignored.